Fix SonarCloud findings: real bugs, blockers, and new-code smells
Real bugs:
- JsoniqParser: the 'validate type EQName { }' branch looked ahead for
  "strict" twice, making the type-name form unparseable (unreachable branch)
- SortedNodeSequence: the dedup guard only skipped ONE duplicate, so runs of
  three or more equal nodes leaked duplicates; the surrounding while never
  looped (unconditional return)
- BooleanValue: zero-argument calls fell through to args[0]
  (ArrayIndexOutOfBoundsException) for any name other than true/false
- ExtractFromDateTime: the per-source switches fell through into the next
  source's cast (ClassCastException for unmatched component/source combos)
- IOUtils.getStringFromFile: a single InputStream.read may return fewer
  bytes than the file length, silently NUL-padding the tail — now reads all
- ProfileOperator: shadowed the parent's id with a second counter, so
  profiler node ids were double-incremented and inconsistent

Interrupt/resource handling:
- Re-interrupt on InterruptedException in the dot-tool runners and
  BlockingParallelizer waits; best-effort temp/spill-file delete() results
  are now checked everywhere (S899/S2142)

Confusion-prone shadowing (blockers):
- ForBind.min/max -> splitMin/splitMax; DefaultCtxItem.type -> itemType

Intentional-exactness suppressions:
- new BigDecimal(double/float) in Dbl/Flt/Cast keeps the EXACT binary value
  (BigDecimal.valueOf would substitute the shortest decimal approximation,
  changing xs:decimal/xs:integer cast results) — annotated with rationale
  instead of "fixed"

Smells in recently-changed code: side-effect-free marker consumption in
GDay/GMD/GMon/Tokenizer, QueryException private-constructor reorder (no
dead flag), yearString if/else ladder, flattened Ordering ternary,
StringBuilder.isEmpty, unnamed catch patterns, pattern instanceof,
Grouping/SumAvgAggregator/ConstructedNodeBuilder complexity extractions,
shared "count" constant, HashMap.newHashMap, unused import, redundant
casts, double-brace init replaced by a named local class.

Full suite 1104/0.

+++ src/main/java/io/brackit/query/compiler/optimizer/cost/OperatorContext.java
@@ -205,7 +205,7 @@ public final class OperatorContext {
     this.inputCardinality = leftCard + rightCard;
     // Ensure output cardinality doesn't overflow
     final double product = (double) leftCard * rightCard * this.selectivity;
-    this.outputCardinality = Math.max(1L, (long) Math.min(product, Long.MAX_VALUE - 1));
+    this.outputCardinality = Math.max(1L, (long) Math.min(product, (double) (Long.MAX_VALUE - 1)));
     return this;
   }
 
@@ -280,7 +280,7 @@ public final class OperatorContext {
     } else {
       // Full unboxing multiplies by array size
       final double product = (double) input * avgArraySize;
-      this.outputCardinality = Math.max(1L, (long) Math.min(product, Long.MAX_VALUE - 1));
+      this.outputCardinality = Math.max(1L, (long) Math.min(product, (double) (Long.MAX_VALUE - 1)));
     }
     return this;
   }
@@ -304,7 +304,7 @@ public final class OperatorContext {
     this.isDescendantDeref = false;
     // Output = input * array_size (unboxing)
     final double product = (double) input * avgArraySize;
-    this.outputCardinality = Math.max(1L, (long) Math.min(product, Long.MAX_VALUE - 1));
+    this.outputCardinality = Math.max(1L, (long) Math.min(product, (double) (Long.MAX_VALUE - 1)));
     return this;
   }
